+++ src/components/Btns/Btn/Btn.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { IBtn } from './interfaces';
+import { IBtn } from '../../interfaces';
 
 export const Btn: React.FC<IBtn> = ({ children, isDisabled }) => (
   <button
