+++ src/components/Btns/DeleteBtn/DeleteBtn.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { IBtnDefault } from './interfaces';
+import { IBtnDefault } from '../../interfaces';
 
 export const DeleteBtn: React.FC<IBtnDefault> = ({ isDisabled }) => (
   <button
